refactor: New modelHandler for model handling
